requirements.txt
streamlit==1.28.0
pandas>=2.3.3
matplotlib>=3.10.6
seaborn>=0.13.2
plotly>=6.3.1
python-dotenv>=1.1.1
openai==2.1.0
scikit-learn>=1.7.2
tiktoken>=0.12.0
tenacity>=8.5.0
numpy>=1.26.4
requests>=2.32.5
fpdf>=1.7.2
